v0.4.11
Networkx package moved to extra packages
Some typo fixes
